chore: moves cms files to folder
